Thu Oct 28 18:58:52 PDT 1999 (KainX)
Basic cleanups on the config stuff.  I made sure that all epplets using the
config subsystem are using atexit(), and I added the same re-prioritizing
code that my epplets have to E-Cpu and E-NetFlame so they wouldn't be as
CPU-intensive.  E-NetFlame has also been upgraded to v0.3 with bandwidth
patches from Shawn (the author), and it also has a config file option for
which interface to monitor (default "eth0") for those who might want to keep
an eye on ppp0.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@1037 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33