The beginning of file magic support and mimetypes ... not really worth playing with it yet, BUT all the functionality needed is basically there ... look in efsd_magic.c for some info -- efsd basically builds a tree of magic tests, like the /usr/share/misc/magic file implicitly does. Lower levels in the tree (assuming root on top) are more specialized, like for mp3's with different bitrates and stuff.
The tests are done in three levels -- first level: filesystem
checks, i.e. when you ask for the mimetype of a directory, you get
something like ext2/dir back (Raster, can you look up the fs magic code
for ext3 and add it???) etc. If it's a regular file, the magic tests
kick in. I put together a small db which already lets me recognize
XCFs and tcpdump traces ... which means string checks and binary
value checks basically work. Really cool. If that fails as well, file
extensions will be checked ... which are not implemented yet.

The whole thing currently is in the client ... not sure if it should
be there or in the server ... but the api is tiny, 3 functions, so I
can move it any time.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@4322 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33