Updated for new debian release.
Cleaned up 'rules' file so that building debs 'raw' out of the CVS will
work, though it will not be identical to the debs uploaded to the main
Debian repository and should *not* be considered official in any way.
They are also not lintian clean.

Added debian/copyright-info which has only the header information for the
COPYING files.  The actual COPYING file must be put in place in the epplets/
directory for a real copyright file to end up in the deb.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@4608 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33