Efsd's filetypes db can now be viewed as html. I've added an XSLT stylesheet that takes the XML file and transforms it into a Javascipt tree widget, using the script from www.treemenu.com (it's BSD licensed etc). Need to send the guy some patches :)
If you want to try it, you'll need a Java XSLT processor, such as Xalan,
and Javascript support enabled in your browser. Go to the doc subdirectory
and do "Make html-filetypes". You'll get the trees in

efsd-manual-0.0.1/filetypes/index.html.

I still need to generate an output file that contains the actual test
details, but that's trivial compared to getting the trees to work.

This was fun :o)


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@5429 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33