* Rewrote ewl_fx.* the old ones were only temporary solutions.    Effects are now pluginable, this is very basic but still powerful.  * /theme/author, /theme/licence, /theme/name keys now exist in the    default theme database.  * Ewl_Table now support object alignments.  * Ewl_Text didn't have entries in the default theme.db, neither    a own thmeme_update call.  * Renamed __ewl_widget_hide to _ewl_widget_hide_fx_clip_box,    since all it did was to hide the fx_clip_box.  * Fixed bug in ewl_config_get_int(). It returned whether or not the    value was fetched properly instead of returning the value.  * Added effects called fade_in, fade_out, and glow.  * Created a Ewl_Widget_Type enum, and added it to Ewl_Widget, mainly    for our fx dispatcher to make use of.  * Changed the FX page in ewl_config to match the new fx code.  * Added a ZERO macro.
i probably forgot to add a whole lot of what i've changed into the changelog,
but at least some of it is there =)
the fxcode still isnt complete, i need to make it possible for themers to
tell what widget will have what effect, and when.
i also need to improve the current code since theres a whole lot i need to
think about before i go any further to make it as powerful as possible.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@5624 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33