Mhmmm okay. I noticed that when I moved some stuff around in libefsd a while ago, I accidentally disabled monitoring of single files. That's fixed now.
But I see very inconsistent behavior when monitoring single files with fam,
both in Efsd and other test apps. Basically, sometimes I just get the
file-exists event for the file, but no change events afterwards. If I
stop the monitor, stop the fam client (again no matter whether it's efsd or
not) and restart it, it usually works. I'll need to check if this is fixed
in the newer fam releases, I'm still on fam 2.6.4.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@5787 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33