Fixed a problem in the detection of XIM support in X11R6. The XIM X11R6 check was being attempted before XIM support was enabled (or disabled) in configure.in. Thus the check was never actually made and X11R6 XIM support was always off.
This caused the resulting Eterm build to have a bug whereby if
the XIM server (kinput2) exits, as it is want to do when
it crashes from time to time, then Eterm would segfault
on reciept of a focus event. By compiling with X11R6 XIM support
Eterm received an event when kinput2 exits and it is all good.

I am not sure if this should be on for all XIM support or not.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@6796 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33