background grabbing evas object, this should make the e-develop list happy
there is a concept here called "freshen", it regrabs the screen with the
coordinates specified.

basically you need tomanually hook in your move,resize settings, OR take
one big snapshot once and pan around at resize time, it's up to you.
This should work fine swallowed into an edje also, but you'll still need
to freshen when the ecore_evas or ecore_x_window is moved/ersized.

This is the same code present in evidence and the iconbar.

See the esmart test for some example usage.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@7827 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33