If you terminate an exe, make sure it really is terminated.
Check it's pulse, hold a mirror to it's mouth, check for brain activity.

Pump another round or two into it, cut off it's head, put a stake through
it's heart.

Sprinkle holy water over it's corpse, hold a black mass to ask Lucifer if
it turned up in hell yet, consult the entrails of rasters three headed
chicken.

If all else fails, complain that the bitch is still alive, and stronger
measures will be needed.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@19475 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33