* fork'n'pipe now has stderr support. * both exe run functions now use the same code. * don't allocate pipes that wan't be used.  This made the code much cleaner. * track and free the exe timers as needed.
E still reports a naughty null timer free at shutdown time, but I don't
think its in ecore_exe.  I'll valgrind it later.

The error fd handler is curently an identical copy of the read fd handler,
with only the names changed.  That's a big slab of code that is duplicated.
I'll merge the two into something generic next.

raster also mentioned that say the first ten lines or so of stderr should
be thrown into a dialog and shown to the user.  I don't know if there is a
way to do that from ecore, or if the user of ecore_exe has to do that
themselves.  The stderr support does line buffered mode just like the read
support, but has not been tested yet.  I'll test properly after the merge.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@19700 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33