* Improve the default theme of Eclair.
This version of eclair will probably be not updated again (I'll maybe add a
filechooser though). Next version of eclair will be an audio player only,
based on a plugin system (like xmms/bmp), it should also have a collection
manager (a mix between the one of itunes and the one of amarok), and will have
a full config interface thanks to Etk.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@22396 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33