Fri Feb 23 18:07:37 2007                        Michael Jennings (mej)
Goodbye new and busted, hello old hotness.

The "new" weather module does not work and makes some unfortunate
design/implementation decisions to boot.  It may be checked in under a
different name ("meteo" perhaps?), but the weather module will remain
as is.  I will maintain it myself if devilhorns doesn't want to.
----------------------------------------------------------------------


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@28455 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33