Use a single XShmImage and XShmPutImage with clip region.
By using a single XShmImage we avoid round trips to X and avoid
having kernel to allocate (and zero) memory on every redraw.

This also enable us to issue a single XShmPutImage() with the whole
XShmImage just by using X Region and setting it as clip on Graphics
Context (GC).

On Nokia N800, expedite gains is about 10fps, while my other test
with fewer objects (and thus drawing areas) I could go from 50fps
to 160fps.

Drawback is that we hold XShmImage until evas is resized or destroyed,
we need a new API to flush engine memory so when it is idle for time
we flush this memory, but it is kept alive during animations.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@30390 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33