Don't leak dbus connection.
We don't need the dbus_connection_ref and _unref because we already
own a reference to the connection after calling dbus_bus_get_private
inside e_dbus_bus_get. Moreover, we're storing a pointer to our
E_DBus_Connection variable in the DBusConnection and setting a
function to free it when the data is replaced or the connection is
finalized (see e_dbus_connection_setup). However, the dbus connection
will never be finalized because we're holding one reference to it
inside our E_DBus_Connection strucuture! This extra reference we're
holding is preventing e_dbus_connection_close from finalizing the dbus
connection (see the Note in the code there). Thus, we're leaking both
the dbus connection and our E_DBus_Connection. Not good, not good at
all. :-/

Patch by Ulisses Furquim, fixing his previous patch.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@31830 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33