Major documentation effort.
This adds "epydoc" compliant documentation for all classes and methods. The
only glitch is that epydoc can't do introspection on pyrex generated classes
and thus cannot process "@param x: desc" instructions, so I've added
"@parm: B{x} desc" as replacement.

In the process of documenting I've come to points that were simpler to
implement than to explain, so two new utilities are born:
 - ClippedSmartObject: a smart object that does almost everything for you,
   just need to implement "resize" virtual.
 - FilledImage: inherits from Image and connects to EVAS_CALLBACK_RESIZE,
   doing the evas_object_image_fill_set() automatically.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@31844 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33