add shrinkable fill flag (bug 178) and fix bug 20
This will slightly change the meaning of the shrink flag:
   'shrink':     the widget wishes to be smaller then its preferred size
   'shrinkable': the widget can be smaller then its preferred size

For more information see the discussion in bug 178.

If you are using the (old) shrink fill policy in your application, you may want
to check if it still does what you want it to. Or if 'shrinkable' isn't better
for the job.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@34803 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33