Helper to make it easy to use Cairo and Evas.
This is a straightforward way to use Cairo as an Evas image object, it
behaves as a double-buffered object, this has pros and cons: if you
want do do animations on Cairo side, it will have some performance
impact since it will render to a memory buffer and then Evas will blit
to visible screen; however if you want to draw a complex Cairo object,
it will be easy to move, resize, colorize and fade it using Evas at
great speed.

Since cairo has no way to notify parts of its surface changed, users
are forced to mark the Evas_Object as dirty themselves.

Just to make it clear: it's not a smart object, it's a regular image
object. Users are encouraged to use Evas functions to change how
images are displayed, set border, fill geometry and more, just need to
take care to not mess with evas_object_image_size_set() or it will
segfault.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@35859 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33