Remove before copy new theme, avoid problems with mmap().
We should not overwrite file contents, what 'cp' does, but instead
create a new file and write it. Due UNIX reference count, the previous
file will continue live untouched in memory as long as E17 uses it. We
could use "install" command, but I fear it's not available everywhere.

This was causing E to freeze after some "make install".



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@35929 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33