Just handle size requests that are larger than 0x0.
This will avoid processing size request signals with 0x0. The problem
was that lots of cases, for example edje_editor's colorpicker dialog,
file dialog and others, were created with size of 0x0, their
wait_size_request was set so the first etk_show() would be ignored,
but then the size_request with 0x0 kicks in and this flag is unset,
then etk_show() runs and show a window of 0x0 size, e17 think it fits
in the most weird places and the result is ugly.

If it breaks something, let me know. Edje Editor works fine, etk_test
seems to work nice too.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@35981 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33