* todo & whites
Ok the the ExchangeSmartObject is ready for testing.
It's an evas smart object declared inside the exchange library, it provide a cool list
of themes for a give category. You can create one with exchange_smart_object_add
and then you can use it as normal evas_object.
The object can act on local themes, on a remote catogory or in a mixed version.
The mixed version is the most powerfull as it check well for update.

Give it a test and report anything, as I want to use this object in the e17
theme & wallpaper configuration.

Hope you like it.

NOTE: to try it out simply INSTALL exchange (and the latest evas) and run
exchange_smart_test.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@37763 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33