shelf ignore mouse_out when going to children windows.
this solves an annoyance with systray: the bounce-on-mouse-over (mouse
over systray would make shelf hide, but after the animation starts the
shelf get mouse-in again and abort the hide, then starts again). But
it will introduce another annoyance: if systray open a menu and you
navigate to it, shelf will not get mouse-out event and thus will not
autohide.

I'm leaving this behavior as it is less annoying. I discussed a bit
with rasterman and solving this is not easy, we'd need to poll for
mouse position and query window tree, this would waste cpu and keep
timeout rate high, draining battery.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@40128 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33