From: Viktor Kojouharov <vkojouharov@gmail.com>
this is a patch for the elm_image so far.

After some valgrind work and with the help of Gustavo, the orient code
for the image is not leak free (according to valgrind). According to
k-s, this way of doing things would avoid any nasty caching errors I've
been getting though valgrind before.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@40314 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33