do not segv on executables that return non-zero.
if an executable returns non-zero a dialog will be present with
information, if it was a .desktop file, then it would use the "Name"
field to name the log and all.

However, the .desktop reference was deleted and the dialog would use a
bogus pointer. Now we take a reference insidethe dialog and just
release it on dialog deletion.

Found by manio, fix by me.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@40987 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33