avoid unwanted clicks on touch-screen.
avoiding the un-wanted clicks with touch-screen devices (especially if
they aren't so precise); it basically allows to set a time delay to be
awaited (pressing over an icon) before sending a "selected" callback.

Thanks to this, it could be needed a longer pressure to select an item
in the file-manager; useful for embedded/touch-based systems.

By: Marco Trevisan (Treviño)



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@41127 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33