get zone from gadcon.
The old code that query the zone based on x,y was causing me
segfaults. When battery module was in a hidden shelf, the position was
outside any zone and then segmentation fault.

As far as I understand, gcc->gadcon->zone should always exist and
always be correct, but I don't have xinerama setup now to confirm
that.

If you know this is wrong, let me know!



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@41147 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33