Added support to creating animated thumbnails from videos.
These thumbnails are generated using an edje template. Each thumbnail is
an edje file, that has a group called "movie/thumb"
The thumbnail can be animated by receiving a signal "animate" or "animate_loop"
(the last one being a loop animation that calls itself when finished), and
the animate_loop program can be stopped by receiving a signal "animate_stop".
The transition time is set based on the time of the animation (that is set
by the call ethumb_video_time_set()).



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@41322 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33