tests: explain some errors messages are expected and check for safety checks.
Many people try "make check" and then complain they have some error
messages, but they are actually expected... but even for the
experienced is hard to know whenever it was expected or not, so
explicitly say that.

Also check if safety checks are enabled or not before doing "break"
tests, otherwise we'll end with segfaults and tests failing. (I have
not tested with them disabled, but should be okay).



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@42360 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33