EDBus: ConnMan service api.
This patch adds the high level "Services" API from ConnMan.

Services is the recommended way to use this system as it will have the
server to do all the low level part, such as dealing with devices,
networks and others.  Network, Devices, Manager and others are still
provided as they might be useful in some cases.

To be done soon is a revamp of econnman module in
e/src/modules/connman. Maybe someone is interested in writing a
fancy elementary application to manage more details?

Patch by: Bruno Dilly



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@44581 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33