connman: update to latest server API.
This API is much more sane and will end in 0.48 (to be released). As
one can see, lots of code is removed.

New missing methods appeared, mainly to avoid IPv4 advanced
configurations. They will be done later.

This commit just remove a bunch of methods, everything related to
'connection' and also renames:

   e_connman_manager_register_agent()
   e_connman_manager_unregister_agent()

to:

   e_connman_manager_agent_register()
   e_connman_manager_agent_unregister()

to be more EFL conformant.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@44797 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33