One more document: user manual of Editje.
This is intented to be newbie-oriented, i.e., people really aware of Edje technology
should also be able to understand Editje minimally and start doing GUI creation with
it.

In the current state, many parts and figures are outdated, we'll be changing them
as development goes by. Anyone is more them welcome to contribute, again.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@46060 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33