Fix segv with shaped windows on GL composite manager.
When we have shaped windows, we must force it to be converted to
non-native so we can modify it later, setting the alpha byte of the
pixels based on the mask.

This fix segv that was happening with Chromium and other shaped
windows as the evas_object_image_data_get() would return NULL for
native surfaces.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@46264 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33