Window/Action/Push in direction is now able to do 2d movements.
Allow one to send window to corners. up-left, up-right, down-left and
down-right are the new possible values.

HINT: for those that don't know, the "Push" has the nice behavior to
do the push in the given direction but stops in the first
window/obstacle in that direction, thus avoiding overlapping windows.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@49433 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33