fix hide/del bug if thumb generation was pending, reuse objects if possible.
As spotted by Brian, we were not cancelling the thumbnail generation
on hide/delete, when the generation finished garbage would be used and
we'd segv. This commit fixes this, also removing the useless
is_generating, as it can be interpreted from wd->id (-1 is not
generating).

As an optimization, we now just delete the view object if it changed
between image and edje (video), otherwise we reuse it.




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@50268 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33