Do not compare pointers to 0
This semantic patch finds places where a pointer is compared to 0. Pointers
should be compared to NULL.

Besides, most of the places should use the E / !E form instead of
E != NULL / E == NULL. This will come in a later patch (badnull.cocci), when
all the corner cases are properly covered.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@50968 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33