Add Elementary Calendar Widget
A calendar is a widget that allows the user to select a date. It has
support to adding marks (holidays and checks by default). The calendar
is displayed month at a time.

Weekdays names and the function used to format month and year to
be displayed can be set, giving more flexibility to this widget.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@51584 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33