Use elm_widget_item common base.
This patch changes all Elm_*_Item to use new infrastructure of
Elm_Widget_Item. This patch is basically a find-and-replace, but even
that was a bit tricky since every code had different names for the
same things, particularly for "view" and "widget" the names ranged
from "o"/"obj" to "base"/"o", quite confusing!

It is worth noting that Genlist and variants that adopted its "item
class" style such as Gengrid and Slideshow already provide a "del()"
callback there and I'm keeping them, however we should reconsider if
it is good to remove them and use elm_widget_item_del_cb_set()
instead. The drawback is that onde would have to set it, while the
other is done automatically when you create the object with the
class. Nonetheless this patch already introduces support to use
elm_widget_item_pre_notify_del(), but it will be void in such code as
nobody ever sets it.

Some new setters and getters were introduced to normalize code... and
they would be required once people write python-bindings for them,
like index and carousel.




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk@52118 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33